refactor(mysql/...): refactor to framework
